Geographically separate Mapbox markers of institutions in same cities
Split records into cities with one or multiple records. Records in
cities that have only one record take the coords of that city.  Records
in cities with multiple records take the location of the institution
(with priority given to coords close to corresponding cities' coords)
with fallback to the city if the institution cannot be resolved.