Make contributors API call return only contributors airtable data
Previously this call will also include the partners data. If in the same session,
the user visits both the acknowledgements page and the partners page, the current
approach will call the Airtable API 3 times, which can be a problem due to
Airtable's low API rate limit. This new approach relies on HTTP caching of the
partners API call to reduce the number of Airtable API calls to 2.